07/11/2023_11:01_ANA

+++ src/main/java/com/sisu/sisu/Dao/EstadoRecetaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.EstadoReceta;
 
-public interface EstadoRecetaDao extends CrudRepository<EstadoReceta, Long> {
+public interface EstadoRecetaDao extends CrudRepository<EstadoReceta, Integer> {
 
 }
