+++ src/main/java/com/sisu/sisu/Service/EstadoCivilServiceImpl.java
@@ -8,36 +8,34 @@ import org.springframework.stereotype.Service;
 import com.sisu.sisu.Dao.EstadoCivilDao;
 import com.sisu.sisu.entitys.TiposEstadoCivil;
 
-
 @Service
-public class EstadoCivilServiceImpl implements EstadoCivilService{
-
+public class EstadoCivilServiceImpl implements EstadoCivilService {
 
     @Autowired
     private EstadoCivilDao estadoCivilDao;
 
     @Override
     public List<TiposEstadoCivil> findAll() {
-       
+
         return (List<TiposEstadoCivil>) estadoCivilDao.findAll();
     }
 
     @Override
     public void save(TiposEstadoCivil estadoCivil) {
-        
+
         estadoCivilDao.save(estadoCivil);
     }
 
     @Override
-    public TiposEstadoCivil findOne(Long id) {
-        
+    public TiposEstadoCivil findOne(Integer id) {
+
         return estadoCivilDao.findById(id).orElse(null);
     }
 
     @Override
-    public void delete(Long id) {
-        
+    public void delete(Integer id) {
+
         estadoCivilDao.deleteById(id);
     }
-    
+
 }
